Wizard: Move queries and states to appropriate places
This moves queries and states based on the component that uses them. This significantly lowers the amount of prop drilling.
